refactor: daily components 폴더명 s 추가
