+++ src/app/daily/_components/Streaks/Streak.styled.tsx
