+++ src/app/daily/_components/Streaks/Streak.tsx
