+++ src/app/daily/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import Calendar from '@/app/daily/_components/Calendar/Calendar';
-import Question from '@/app/daily/_components/Question/Question';
-import Streak from '@/app/daily/_components/Streak/Streak';
+import Question from '@/app/daily/_components/Questions/Question';
+import Streak from '@/app/daily/_components/Streaks/Streak';
 import Breadcrumb from '@/components/molecules/Breadcrumb/Breadcrumb';
 import useGetDailyQuery from '@/hooks/api/daily/useGetDaily';
 import useGetStreakDaysQuery from '@/hooks/api/streak/useGetStreakDays';
